My portion of the .csv script, the DCB.
(This is where I'll be pushing stuff for this specfically)
It reads the DCB sheet from the database, with the goal in mind of making it robust and easy to modify. 

It uses regex to identify DCB boards, and then sorts them into three dictionaries (assembled, not assembled, unknown) with a linear algorithmic complexity. The dictionaries have keys based on the ID numbers, and a value with the relevant information from each row of the database split into a list of Strings. It'd designed to be modified later so the values from each line can further split as necessary.